+++ src/main/kotlin/com/wafflytime/user/auth/api/AuthController.kt
@@ -1,12 +1,12 @@
-package com.wafflytime.user.auth.controller.api
+package com.wafflytime.user.auth.api
 
 import com.wafflytime.config.ExemptAuthentication
 import com.wafflytime.config.ExemptEmailVerification
 import com.wafflytime.config.UserIdFromToken
-import com.wafflytime.user.auth.controller.dto.AuthToken
+import com.wafflytime.user.auth.api.dto.AuthToken
 import com.wafflytime.user.auth.service.AuthTokenService
-import com.wafflytime.user.auth.controller.dto.LoginRequest
-import com.wafflytime.user.auth.controller.dto.SignUpRequest
+import com.wafflytime.user.auth.api.dto.LoginRequest
+import com.wafflytime.user.auth.api.dto.SignUpRequest
 import com.wafflytime.user.auth.service.LocalAuthService
 import jakarta.validation.Valid
 import org.springframework.web.bind.annotation.DeleteMapping
