+++ src/main/kotlin/com/wafflytime/user/auth/api/dto/AuthToken.kt
@@ -1,3 +1,3 @@
-package com.wafflytime.user.auth.controller.dto
+package com.wafflytime.user.auth.api.dto
 
 data class AuthToken(val accessToken: String, val refreshToken: String)
