+++ src/main/kotlin/com/wafflytime/user/auth/api/dto/LoginRequest.kt
@@ -1,8 +1,8 @@
-package com.wafflytime.user.auth.controller.dto
+package com.wafflytime.user.auth.api.dto
 
 import jakarta.validation.constraints.NotBlank
 
-data class SignUpRequest(
+data class LoginRequest(
     @field:NotBlank
     val id: String,
     @field:NotBlank
