+++ src/main/kotlin/com/wafflytime/user/auth/api/dto/OAuthToken.kt
@@ -1,4 +1,4 @@
-package com.wafflytime.user.auth.controller.dto
+package com.wafflytime.user.auth.api.dto
 
 import com.fasterxml.jackson.annotation.JsonProperty
