+++ src/main/kotlin/com/wafflytime/user/auth/api/dto/SignUpRequest.kt
@@ -1,8 +1,8 @@
-package com.wafflytime.user.auth.controller.dto
+package com.wafflytime.user.auth.api.dto
 
 import jakarta.validation.constraints.NotBlank
 
-data class LoginRequest(
+data class SignUpRequest(
     @field:NotBlank
     val id: String,
     @field:NotBlank
