--- src/main/kotlin/com/wafflytime/user/auth/controller/dto/SocialLoginRequest.kt
@@ -1,8 +1,0 @@
-package com.wafflytime.user.auth.controller.dto
-
-import jakarta.validation.constraints.NotBlank
-
-data class SocialLoginRequest(
-    @field:NotBlank
-    val socialEmail: String,
-)
