--- src/main/kotlin/com/wafflytime/user/auth/controller/dto/SocialSignUpRequest.kt
@@ -1,10 +1,0 @@
-package com.wafflytime.user.auth.controller.dto
-
-import jakarta.validation.constraints.NotBlank
-
-data class SocialSignUpRequest(
-    @field:NotBlank
-    val socialEmail: String,
-    @field:NotBlank
-    val univEmail: String,
-)
